Split AbstractArrays and ConcreteArrays classes into files
