+++ src/scene/ASBuildScratchpad.hpp
@@ -19,7 +19,8 @@
 #include <optix_stubs.h>
 
 #include <Optix.hpp>
-#include <memory/ConcreteArrays.hpp>
+#include <memory/DeviceSyncArray.hpp>
+#include <memory/HostPinnedArray.hpp>
 
 /**
  * Helper class to manage buffers used for building acceleration (GAS, IAS) structures and perform their compaction.
